Fit-Track-UI: Files Convention Updated
