+++ fit-track-ui/react/src/components/Common/Modals/workout-log-modal/workout-log-modal.scss
