+++ fit-track-ui/react/src/components/Common/Modals/workout-modal/workout-modal.scss
