+++ fit-track-ui/react/src/components/Common/Widgets/health-info/health-widget.scss
