+++ fit-track-ui/react/src/components/Common/Widgets/profile-details/profile-widget.scss
