+++ fit-track-ui/react/src/components/Common/Widgets/workout-history/workout-history-widget.scss
