+++ fit-track-ui/react/src/components/Features/Logs/Logs.tsx
@@ -1,10 +1,10 @@
 import React, {useEffect, useState} from "react";
-import Sidebar from "../Layout/Sidebar/Sidebar.tsx";
-import Navbar from "../Layout/Navbar/Navbar.tsx";
-import {getAllWorkoutsByCustomerId, getCustomer} from "../../services/client.ts";
+import Sidebar from "../../Layout/Sidebar/Sidebar.tsx";
+import Navbar from "../../Layout/Navbar/Navbar.tsx";
+import {getAllWorkoutsByCustomerId, getCustomer} from "../../../services/client.ts";
 import "./logs.scss";
-import {Customer, Workout} from "../../types/index.ts";
-import WorkoutLogModal from "../Common/Modals/workout-log-modal/WorkoutLogModal.tsx";
+import {Customer, Workout} from "../../../types/index.ts";
+import WorkoutLogModal from "../../Common/Modals/workout-log-modal/WorkoutLogModal.tsx";
 
 const Logs = () => {
     const [customer, setCustomer] = useState<Customer>({});
