+++ fit-track-ui/react/src/components/Layout/Navbar/navbar.scss
