+++ fit-track-ui/react/src/components/Layout/Sidebar/sidebar.scss
