+++ fit-track-ui/react/src/shared/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import {useEffect} from "react";
 import {useNavigate} from "react-router-dom";
-import {useAuth} from "../../context/AuthContext";
+import {useAuth} from "../context/AuthContext";
 
 const ProtectedRoute = ({children}: { children: any }) => {
